clean, stage 1
cleaned almost all unused code, also refactor a bit to merge exact code
