+++ OpenBible/Managing/VerseManager.swift
@@ -139,37 +139,6 @@ class VerseManager: Manager {
 
 extension VerseManager {
     func setVerses(from strArray: [String]) {
-        var verseRanges = [Range<Int>]()
-        var pendingRange: Range<Int>? = nil
-        for verse in strArray {
-            if !("0"..."9" ~= verse[0]) {
-                if let v = Int(verse[verse.index(after: verse.startIndex)...]) {
-                    switch verse[0] {
-                    case "-":
-                        if pendingRange != nil {
-                            pendingRange = Range(uncheckedBounds: (pendingRange!.lowerBound, v + 1))
-                        } else {
-                            pendingRange = Range(uncheckedBounds: (v, v + 1))
-                        }
-                    case ",",".":
-                        if pendingRange != nil {
-                            verseRanges.append(pendingRange!)
-                        }
-                        pendingRange = Range(uncheckedBounds: (v, v + 1))
-                    default:break
-                    }
-                }
-            } else {
-                let v = Int(verse)!
-                if pendingRange != nil {
-                    verseRanges.append(pendingRange!)
-                }
-                pendingRange = Range(uncheckedBounds: (v,v + 1))
-            }
-        }
-        if pendingRange != nil {
-            verseRanges.append(pendingRange!)
-        }
-        versesRanges = verseRanges
+        versesRanges = getVerseRanges(from: strArray)
     }
 }
