--- OpenBible/Supplement/Progress/Progressable.h
@@ -1,19 +1,0 @@
-//
-//  Progressable.h
-//  Progressable
-//
-//  Created by Martin MOIZARD-LANVIN on 27/09/2016.
-//  Copyright © 2016 Martin Moizard. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for Progressable.
-FOUNDATION_EXPORT double ProgressableVersionNumber;
-
-//! Project version string for Progressable.
-FOUNDATION_EXPORT const unsigned char ProgressableVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <Progressable/PublicHeader.h>
-
-
