+++ macB.xcworkspace/xcuserdata/deniz.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -73,22 +73,6 @@
             shouldBeEnabled = "No"
             ignoreCount = "0"
             continueAfterRunningActions = "No"
-            filePath = "macB/AppDelegate.swift"
-            timestampString = "571918349.464983"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "120"
-            endingLineNumber = "120"
-            landmarkName = "rewriteSharingObjects(_:)"
-            landmarkType = "7">
-         </BreakpointContent>
-      </BreakpointProxy>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            shouldBeEnabled = "No"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
             filePath = "macB/Managing/SharingManager.swift"
             timestampString = "572471106.425981"
             startingColumnNumber = "9223372036854775807"
