+++ macB/Supplement/Protocols.swift
@@ -8,18 +8,6 @@
 
 import Foundation
 
-protocol DownloadDelegate {
-    func initiateDownload(by key: String, completition: ((Bool) -> Void)?)
-    func initiateRemoval(by key: String, completition: ((Bool) -> Void)?)
-    func initiateRemoval(by index: Int, completition: ((Bool) -> Void)?)
-}
-
-extension DownloadDelegate {
-    func initiateDownload(by key: String, completition: ((Bool) -> Void)? = nil) {}
-    func initiateRemoval(by key: String, completition: ((Bool) -> Void)? = nil) {}
-    func initiateRemoval(by index: Int, completition: ((Bool) -> Void)? = nil) {}
-}
-
 protocol ModelUpdateDelegate {
     var hashValue: Int {get}
     func modelChanged()
@@ -52,17 +40,6 @@ protocol SplitViewParticipant {
     var hashValue: Int {get}
     func splitViewParticipantDidEndScrolling()
     func splitViewParticipantDidScrolled(to offsetRatio: CGFloat)
-}
-
-protocol SharingSelectingDelegate {
-    func sharingObjectWasSelected(with status: Bool, being: Int)
-}
-
-protocol BonjourManagerDelegate {
-    func bonjourDidChanged(isConnected: Bool, to host: String?, at port: Int?)
-    func bonjourServiceUpdated(to status: String)
-    func bonjourDidRead(message: String?)
-    func bonjourDidWrite()
 }
 
 protocol SideMenuDelegate {
